requirements.txt
matplotlib>=3.2.2
numpy>=1.20
Pillow>=7.1.2
torch==1.9.0
torchvision==0.10.0
tqdm>=4.41.0
tensorboard>=2.4.1